1. Убран форматтер. 2. Упрощен импорт файлов
