+++ NewPlatform.Flexberry.Analytics/NewPlatform.Flexberry.Analytics.sln
@@ -1,15 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26228.9
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28922.388
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Abstractions", "Abstractions\Abstractions.csproj", "{A4850BD2-3CE6-4BB1-B0AD-7E26A9AE7F16}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReportManager", "PentahoReport\ReportManager.csproj", "{9E941213-D6A7-4063-B77F-FA2CD6B58F14}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Report_WebApi", "Report_WebApi\Report_WebApi.csproj", "{0DA19C15-A104-4CED-A4C4-22D9B129AB1B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Report_Formatters", "Report_Formatters\Report_Formatters.csproj", "{0D9B2C84-F7E8-422D-BBA6-1E27F94F057C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReportWebApiService", "ReportWebApiService\ReportWebApiService.csproj", "{EBDA6828-35E2-4C0E-90EA-1DA1ADFA2B24}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,19 +23,16 @@ Global
 		{9E941213-D6A7-4063-B77F-FA2CD6B58F14}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9E941213-D6A7-4063-B77F-FA2CD6B58F14}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9E941213-D6A7-4063-B77F-FA2CD6B58F14}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0DA19C15-A104-4CED-A4C4-22D9B129AB1B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0DA19C15-A104-4CED-A4C4-22D9B129AB1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0DA19C15-A104-4CED-A4C4-22D9B129AB1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0DA19C15-A104-4CED-A4C4-22D9B129AB1B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0D9B2C84-F7E8-422D-BBA6-1E27F94F057C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0D9B2C84-F7E8-422D-BBA6-1E27F94F057C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0D9B2C84-F7E8-422D-BBA6-1E27F94F057C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0D9B2C84-F7E8-422D-BBA6-1E27F94F057C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EBDA6828-35E2-4C0E-90EA-1DA1ADFA2B24}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EBDA6828-35E2-4C0E-90EA-1DA1ADFA2B24}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EBDA6828-35E2-4C0E-90EA-1DA1ADFA2B24}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EBDA6828-35E2-4C0E-90EA-1DA1ADFA2B24}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0BE8CBBC-0C08-48B4-8A22-CDA83B509B0B}
 		EnterpriseLibraryConfigurationToolBinariesPath = packages\Unity.2.1.505.2\lib\NET35;packages\Unity.Interception.2.1.505.0\lib\NET35
 	EndGlobalSection
 EndGlobal
