+++ NewPlatform.Flexberry.Analytics/ReportWebApiService/App_Start/WebApiConfig.cs
@@ -1,0 +1,29 @@
+﻿using NewPlatform.Flexberry.AspNet.WebApi.Cors;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web.Http;
+using Unity;
+using Unity.AspNet.WebApi;
+
+namespace ReportWebApiService
+{
+    public static class WebApiConfig
+    {
+        public static void Register(HttpConfiguration config)
+        {
+            // Конфигурация и службы веб-API
+  
+            config.EnableCors(new DynamicCorsPolicyProvider(true));
+
+            // Маршруты веб-API
+            config.MapHttpAttributeRoutes();
+
+            config.Routes.MapHttpRoute(
+                name: "DefaultApi",
+                routeTemplate: "api/{controller}/{id}",
+                defaults: new { id = RouteParameter.Optional }
+            );
+        }
+    }
+}
