+++ NewPlatform.Flexberry.Analytics/ReportWebApiService/Controllers/ReportController.cs
@@ -1,19 +1,12 @@
-﻿namespace Report_WebApi.Controllers
+﻿namespace ReportWebApiService.Controllers
 {
     using Abstractions;
     using ICSSoft.STORMNET;
     using Newtonsoft.Json.Linq;
     using System;
-    using System.Net;
-    using System.Net.Http;
-    using System.Net.Http.Headers;
     using System.Threading;
     using System.Threading.Tasks;
     using System.Web.Http;
-    using System.Web.Http.Cors;
-
-
-    //[OverrideAuthorization]
 
     [RoutePrefix("api/Report")]
     public class ReportController : ApiController
