+++ NewPlatform.Flexberry.Analytics/ReportWebApiService/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="ReportWebApiService.WebApiApplication" Language="C#" %>
