+++ NewPlatform.Flexberry.Analytics/ReportWebApiService/Global.asax.cs
@@ -1,0 +1,12 @@
+using System.Web.Http;
+
+namespace ReportWebApiService
+{
+    public class WebApiApplication : System.Web.HttpApplication
+    {
+        protected void Application_Start()
+        {
+            GlobalConfiguration.Configure(WebApiConfig.Register);
+        }
+    }
+}
