+++ NewPlatform.Flexberry.Analytics/ReportWebApiService/Web.Debug.config
