+++ NewPlatform.Flexberry.Analytics/ReportWebApiService/Web.Release.config
