+++ NewPlatform.Flexberry.Analytics/ReportWebApiService/packages.config
@@ -1,0 +1,22 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="log4net" version="2.0.3" targetFramework="net45" />
+  <package id="Microsoft.AspNet.Cors" version="5.2.3" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi" version="5.2.7" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Client.ru" version="5.2.7" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Core.ru" version="5.2.7" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Cors" version="5.2.3" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.7" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.WebHost.ru" version="5.2.7" targetFramework="net45" />
+  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="2.0.0" targetFramework="net45" />
+  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net45" />
+  <package id="NewPlatform.Flexberry.AspNet.WebApi.Cors" version="1.1.1" targetFramework="net45" />
+  <package id="NewPlatform.Flexberry.LogService" version="1.1.0" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net45" />
+  <package id="Unity.Abstractions" version="3.3.0" targetFramework="net45" />
+  <package id="Unity.AspNet.WebApi" version="5.0.13" targetFramework="net452" />
+  <package id="Unity.Container" version="5.8.6" targetFramework="net45" />
+  <package id="WebActivatorEx" version="2.2.0" targetFramework="net45" />
+</packages>
