--- NewPlatform.Flexberry.Analytics/Report_Formatters/packages.config
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ClosedXML" version="0.93.1" targetFramework="net452" />
-  <package id="DocumentFormat.OpenXml" version="2.7.2" targetFramework="net452" />
-  <package id="ExcelNumberFormat" version="1.0.3" targetFramework="net452" />
-  <package id="FastMember.Signed" version="1.3.0" targetFramework="net452" />
-  <package id="Microsoft.AspNet.OData" version="5.7.0" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.2" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.2" targetFramework="net452" />
-  <package id="Microsoft.OData.Core" version="6.13.0" targetFramework="net452" />
-  <package id="Microsoft.OData.Edm" version="6.13.0" targetFramework="net452" />
-  <package id="Microsoft.Spatial" version="6.13.0" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net452" />
-  <package id="System.Net.Http" version="2.0.20126.16343" targetFramework="net452" />
-  <package id="System.Net.Http.Formatting.Extension" version="5.2.3.0" targetFramework="net452" />
-</packages>
