--- NewPlatform.Flexberry.Analytics/Report_WebApi/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Report_WebApi.WebApiApplication" Language="C#" %>
