--- NewPlatform.Flexberry.Analytics/Report_WebApi/packages.config
@@ -1,30 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="CommonServiceLocator" version="1.0" targetFramework="net452" />
-  <package id="log4net" version="2.0.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.Cors" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Client.ru" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Core.ru" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Cors" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.Extensions.Compression.Server" version="2.0.6" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi.WebHost.ru" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.3" targetFramework="net452" />
-  <package id="Microsoft.IO.RecyclableMemoryStream" version="1.1.0.0" targetFramework="net452" />
-  <package id="Microsoft.Net.Compilers" version="2.4.0" targetFramework="net452" developmentDependency="true" />
-  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
-  <package id="NewPlatform.Flexberry.AspNet.WebApi.Cors" version="1.1.1" targetFramework="net452" />
-  <package id="NewPlatform.Flexberry.LogService" version="1.1.0" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net452" />
-  <package id="SharpZipLib" version="0.86.0" targetFramework="net452" />
-  <package id="Unity" version="5.0.2" targetFramework="net452" />
-  <package id="Unity.Abstractions" version="3.3.1" targetFramework="net452" />
-  <package id="Unity.AspNet.WebApi" version="5.0.13" targetFramework="net452" />
-  <package id="Unity.Configuration" version="5.2.5" targetFramework="net452" />
-  <package id="Unity.Container" version="5.8.11" targetFramework="net452" />
-  <package id="Unity.Interception" version="2.1.505.0" targetFramework="net452" />
-  <package id="WebActivatorEx" version="2.2.0" targetFramework="net452" />
-</packages>
